Add horizontal scrolling to code blocks
Adding vertical padding and horizontal scrolling to code blocks. The
padding makes it easier to scroll single line code blocks and the
scrolling helps so that readers can scroll the block while still
keeping text in focus. I still prefer being able to scroll anywhere
on the screen, but it seems the general feel is that it's annoying
to have horizontal scroll when the code block is far away.