Increase max-width
I saw this site https://steveklabnik.com/writing/too-many-words-about-rusts-function-syntax
and really liked the width. I'm tired of fighting for standard narrow
columns, a bit wider on desktop is good.
I wanted to use 80ch (because it's less characters), but the header and
the article both use the width, and because of font-size differences
80ch isn't the same for the header and the body