Remove footnotes
They didn't work with reader mode on firefox and I realized that they're
unnecessary. I was either using them as hyperlinks (which are better
inline [1 click instead of 2]) or as sidenotes, which should be in the
sidebar. If I want to do academic paper style I should be releasing in
pdf.